completed part 4

+++ part4/blog_list/controllers/blogs.js
@@ -1,13 +1,22 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
+// Fetch all blogs (public access)
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
+// Middleware to protect routes below
+blogsRouter.use(middleware.tokenExtractor)
+blogsRouter.use(middleware.userExtractor)
+
+
+// Add new blog (if user is authorised)
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
 
   if (!body.title || !body.url) {
     return response.status(400).json({ error: 'title or url missing' });
@@ -17,22 +26,41 @@ blogsRouter.post('/', async (request, response) => {
     title: body.title,
     author: body.author,
     likes: body.likes,
-    url: body.url
+    url: body.url,
+    user: user._id
   })
   
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
+// Delete a blog (if user is authorised)
 blogsRouter.delete('/:id', async (request, response) => {
-  const result = await Blog.findByIdAndDelete(request.params.id);
+  const user = request.user
+  console.log('user:', user)
+
+  const blogToDelete = await Blog.findById(request.params.id).populate('user')
+  console.log('blogToDelete:', blogToDelete)
+  if (!blogToDelete) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blogToDelete.user._id.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'blogs can only be deleted by their creators' });
+  } 
+
+  const result = await Blog.findByIdAndDelete(request.params.id)
+  console.log('result of deletion:', result)
   if (result) {
-    response.status(204).end();
+    response.status(204).end()
   } else {
-    response.status(404).json({ error: 'blog not found' });
+    response.status(500).json({ error: 'blog deletion failed' })
   }
 })
 
+// Update a pre-existing blog's likes
 blogsRouter.put('/:id', async (request, response) => {
   const body = await request.body
